fix: removes the require loop on navigators definition
