+++ ts/theme/navigations.ts
@@ -1,0 +1,26 @@
+import {
+  IOColors,
+  IOThemeDark,
+  IOThemeLight
+} from "@pagopa/io-app-design-system";
+import { DarkTheme, DefaultTheme } from "@react-navigation/native";
+
+// React Navigation Themes
+// Dark & Light mode
+export const IONavigationDarkTheme = {
+  ...DarkTheme,
+  colors: {
+    ...DarkTheme.colors,
+    background: IOColors[IOThemeDark["appBackground-primary"]],
+    card: IOColors[IOThemeDark["appBackground-primary"]]
+  }
+};
+
+export const IONavigationLightTheme = {
+  ...DefaultTheme,
+  colors: {
+    ...DefaultTheme.colors,
+    background: IOColors[IOThemeLight["appBackground-primary"]],
+    card: IOColors[IOThemeLight["appBackground-primary"]]
+  }
+};
